Move ACP build output to ignored dist
