+++ mastra-agents/package.json
@@ -7,12 +7,12 @@
     "node": ">=22.13.0"
   },
   "scripts": {
-    "acp": "node src/acp/stdio.js",
-    "acp:build": "tsc --ignoreConfig --target ESNext --module NodeNext --moduleResolution NodeNext --strict --skipLibCheck --types node --rootDir acp --outDir src/acp --declaration acp/*.ts",
-    "acp:smoke": "node src/acp/smoke-client.js",
+    "acp": "npm run acp:build && node dist/acp/stdio.js",
+    "acp:build": "tsc --ignoreConfig --target ESNext --module NodeNext --moduleResolution NodeNext --strict --skipLibCheck --types node --rootDir acp --outDir dist/acp --declaration acp/*.ts",
+    "acp:smoke": "npm run acp:build && node dist/acp/smoke-client.js",
     "dev": "node scripts/with-env.mjs mastra dev",
     "build": "node scripts/with-env.mjs mastra build",
-    "test": "node --test \"src/**/*.test.js\" \"test/**/*.test.js\"",
+    "test": "npm run acp:build && node --test \"src/**/*.test.js\" \"test/**/*.test.js\"",
     "start": "node scripts/with-env.mjs mastra start",
     "studio": "node scripts/with-env.mjs node scripts/run-studio.mjs"
   },
@@ -34,7 +34,7 @@
     "zod": "3.25.76"
   },
   "bin": {
-    "mastra-agents-acp": "src/acp/stdio.js"
+    "mastra-agents-acp": "dist/acp/stdio.js"
   },
   "devDependencies": {
     "@types/node": "^25.6.0",
