--- mastra-agents/src/acp/adapter.d.ts
@@ -1,3 +1,0 @@
-import type { AgentSideConnection } from '@agentclientprotocol/sdk';
-import type { ACPAgent, MastraAcpAgentOptions } from './types.js';
-export declare function createMastraAcpAgentHandler(conn: AgentSideConnection, options: MastraAcpAgentOptions): ACPAgent;
