--- mastra-agents/src/acp/event-mapper.d.ts
@@ -1,3 +1,0 @@
-import type { SessionUpdate, ToolCallUpdate } from '@agentclientprotocol/sdk';
-export declare function inferToolKind(name?: string): ToolCallUpdate['kind'];
-export declare function mapMastraChunkToUpdates(chunk: unknown): SessionUpdate[];
