--- mastra-agents/src/acp/index.js
@@ -1,6 +1,0 @@
-import { createMastraAcpAgentHandler } from './adapter.js';
-export * from './types.js';
-export * from './adapter.js';
-export function createMastraAcpAgent(options) {
-    return (conn) => createMastraAcpAgentHandler(conn, options);
-}
