--- mastra-agents/src/acp/mastra-stream.d.ts
@@ -1,1 +1,0 @@
-export declare function streamMastraAgent(baseUrl: string | undefined, agentId: string, payload: unknown, signal?: AbortSignal): AsyncGenerator<unknown>;
