--- mastra-agents/src/acp/session-store.d.ts
@@ -1,16 +1,0 @@
-import type { MastraAcpSession } from './types.js';
-export declare class MastraAcpSessionStore {
-    private readonly sessions;
-    create(params: {
-        sessionId?: string;
-        agentId: string;
-        cwd: string;
-        resourceId?: string;
-        threadId?: string;
-        defaultModeId: string;
-        defaultModelId: string;
-    }): MastraAcpSession;
-    get(sessionId: string): MastraAcpSession | undefined;
-    update(session: MastraAcpSession): void;
-    delete(sessionId: string): void;
-}
