--- mastra-agents/src/acp/slash-commands.d.ts
@@ -1,2 +1,0 @@
-import type { AvailableCommand } from '@agentclientprotocol/sdk';
-export declare function getSlashCommands(): AvailableCommand[];
