--- mastra-agents/src/acp/smoke-client.d.ts
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-export {};
