--- mastra-agents/src/acp/stdio.d.ts
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-export {};
