--- mastra-agents/src/acp/types.js
@@ -1,1 +1,0 @@
-export {};
