+++ mastra-agents/src/index.ts
@@ -12,6 +12,6 @@ export * from './tools/index.js';
 export * from './workflows/index.js';
 export * from './scorers/index.js';
 export * from './daytona/index.js';
-export * from './acp/index.js';
+export * from '../acp/index.js';
 export { mastra } from './mastra/index.js';
 export { workspace, workspaceAccessRoots, workspaceCommandCwd, workspaceRoot } from './workspace.js';
